Move app_data.rs to eframe crate since it is specific to GUI framework
