+++ eframe/src/lib.rs
@@ -1,5 +1,6 @@
 #![warn(clippy::all, rust_2018_idioms)]
 
 mod app;
+mod app_data;
 mod bg_image;
 pub use app::TemplateApp;
